Fix error in scalar type, fixes issue #29
The scalar type is documented to act on signed integer values, but the
formatting was erroneously treating the data as unsigned.

+++ server/types.c
@@ -274,7 +274,8 @@ static error__t scalar_format(
     unsigned int value, char string[], size_t length)
 {
     struct scalar_state *state = type_data;
-    return format_double(string, length, state->scale * value + state->offset);
+    return format_double(string, length,
+        state->scale * (int) value + state->offset);
 }
